Reworked project structure. Added custom M2M Relationship between Users and Events
