+++ user_handler/apps.py
@@ -1,6 +1,6 @@
 from django.apps import AppConfig
 
 
-class EventssiteConfig(AppConfig):
+class UserHandlerConfig(AppConfig):
     default_auto_field = 'django.db.models.BigAutoField'
-    name = 'eventsSite'
+    name = 'user_handler'
